refactor: 태그 엣지, 릴레이션 디렉토리 구조 변경
