+++ src/main/java/com/example/oatnote/memotag/service/tag/edge/TagEdgeRepository.java
@@ -1,10 +1,10 @@
-package com.example.oatnote.memotag.service.tagEdge;
+package com.example.oatnote.memotag.service.tag.edge;
 
 import java.util.Optional;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import com.example.oatnote.memotag.service.tagEdge.model.TagEdge;
+import com.example.oatnote.memotag.service.tag.edge.model.TagEdge;
 
 public interface TagEdgeRepository extends MongoRepository<TagEdge, String> {
