+++ src/main/java/com/example/oatnote/memotag/service/tag/relation/exception/TagsRelationNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.memotag.service.tagsRelation.exception;
+package com.example.oatnote.memotag.service.tag.relation.exception;
 
 public class TagsRelationNotFoundException extends RuntimeException {
     public TagsRelationNotFoundException(Object p0) {
